+++ app/src/main/java/com/kylecorry/trail_sense/weather/domain/LowPassFilter.kt
@@ -1,10 +1,10 @@
 package com.kylecorry.trail_sense.weather.domain
 
-class LowPassFilter(private val alpha: Double, initialValue: Double = 0.0) {
+class LowPassFilter(private val alpha: Float, initialValue: Float = 0f) {
 
     private var estimate = initialValue
 
-    fun filter(measurement: Double): Double {
+    fun filter(measurement: Float): Float {
         estimate = (1 - alpha) * estimate + alpha * measurement
         return estimate
     }
